Added new MaunWindow class which open a new window, deleted opening window from logic folder.
