+++ src/main/java/studos/gui/MainWindow.java
@@ -1,0 +1,46 @@
+/**
+ * Info about this package doing something for package-info.java file.
+ */
+
+package studos.gui;
+
+import javafx.event.ActionEvent;
+import javafx.event.EventHandler;
+import javafx.scene.Scene;
+import javafx.scene.control.Button;
+import javafx.scene.layout.StackPane;
+import javafx.stage.Stage;
+
+/**
+ * Front end for main window.
+ */
+
+public class MainWindow {
+
+    /**
+     * It's fornt method for new window.
+     *
+     */
+
+    public void mainWindow() {
+        final Stage primaryStage = new Stage();
+        primaryStage.setTitle("Hello World!");
+        final Button btn = new Button();
+        btn.setText("Say 'Hello World'");
+        btn.setOnAction(new EventHandler<ActionEvent>() {
+
+            @Override
+            public void handle(final ActionEvent event) {
+                System.out.println("Hello World!");
+            }
+        });
+        final int windowHeight = 300;
+        final int windowWidth = 250;
+
+        final StackPane root = new StackPane();
+        root.getChildren().add(btn);
+        primaryStage.setScene(new Scene(root, windowHeight, windowWidth));
+        primaryStage.show();
+    }
+}
+
